services/leaf-quiz-service/requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Pydantic and validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
PyJWT==2.8.0

# HTTP client
httpx==0.25.2

# Redis
redis==5.0.1

# --- HuggingFace API Section ---
# Use HuggingFace Inference API instead of local models
requests==2.31.0

# Text processing
nltk==3.8.1
spacy==3.7.2

# Utilities
numpy
pandas
scikit-learn

# Environment and configuration
python-dotenv==1.0.0

# Logging
structlog==23.2.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1 